Move nl_diff to a more convenient location
