+++ pyomo/repn/tests/ampl/test_ampl_nl.py
@@ -28,7 +28,7 @@ from pyomo.environ import (
     ExternalFunction,
     value,
 )
-from .nl_diff import load_and_compare_nl_baseline
+from ..nl_diff import load_and_compare_nl_baseline
 
 import pyomo.repn.plugins.ampl.ampl_ as ampl_
 import pyomo.repn.plugins.nl_writer as nl_writer
