+++ pyomo/repn/tests/ampl/test_suffixes.py
@@ -29,7 +29,7 @@ from pyomo.environ import (
     SOSConstraint,
     sum_product,
 )
-from .nl_diff import load_and_compare_nl_baseline
+from ..nl_diff import load_and_compare_nl_baseline
 
 currdir = this_file_dir()
